fix(irc): after a server forces a nick-change, the bridge loses track of the puppet
The server forces a nick-change when someone is claiming
their existing account. Or in some rare cases during a net-split.

When this happens, try to claim the original name back, with the
fallthoughs as normal.

Funny enough, if we do this too quickly, we reclaim the nick
that we lost in case someone is claiming the nick. So be nice,
and wait a bit of time before we actually do that.