feat(irc): disconnect IRC puppets 3 hours after the Discord user went offline
When someone says something in the IRC bridge channel on Discord,
they will get an IRC presence. This IRC puppet will remain connected
until

- 3 hours after the Discord user went offline/invisible.

OR

- 3 hours after the Discord user said anything in the IRC bridged
  channel after going invisible.

Which ever comes last. This three hour grace period is to comply
more with the usage of Discord on mobile. Users tend to come online,
say something, go offline. And repeat this a few time during the
conversation.