feat(irc): disconnect IRC puppets a while after the Discord user went offline
When a Discord user is marked as offline, they will disconnect from
IRC after a configurable timeout. Additionally, as offline can also
mean the user is invisible, this countdown is reset every time a
user talks while being offline.

Depending on your network, the right timeout might vary. It is
advisible to use a long time, like more than a day. This works
best for users that want to be online on IRC all day every day,
and still have a decent night sleep. Additionally, for people that
don't actually care, also no harm is done.

Because of this, a default value of 48 hours is set.