feat(irc): on kill, don't reconnect (#138)
When a puppet is killed on IRC, it means an oper stepped in and
didn't appreciate what the puppet was saying. Don't annoy the
IRC oper by reconnecting.

This is only really used in cases of spambots; although they are
banned on Discord really quick, IRC doesn't see that, and they
rightfully take their own action. So respond to that in a way
that indicates we acknowledged the request.